+++ native/c/zowex.cpp
@@ -140,6 +140,9 @@ int main(int argc, char *argv[])
                              make_aliases("--wait"),
                              "wait for responses", ArgType_Flag, false,
                              ArgValue(true));
+  issue_cmd->add_keyword_arg("timeout",
+                             make_aliases("--timeout"),
+                             "timeout in seconds", ArgType_Single, false);
   issue_cmd->add_positional_arg("command", "command to run, e.g. 'D IPLINFO'",
                                 ArgType_Single, true);
   issue_cmd->set_handler(handle_console_issue);
@@ -577,9 +580,15 @@ int handle_console_issue(const ParseResult &result)
   ZCN zcn = {0};
 
   string console_name = result.find_kw_arg_string("console-name");
-  cout << "console name " << console_name << endl;
+  int timeout = result.find_kw_arg_int("timeout");
+
   string command = result.find_pos_arg_string("command");
   bool wait = result.find_kw_arg_bool("wait");
+
+  if (timeout > 0)
+  {
+    zcn.timeout = timeout;
+  }
 
   rc = zcn_activate(&zcn, console_name);
   if (0 != rc)